+++ python/sdk/setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = prelude-sdk
-version = 2.5.8
+version = 2.5.9
 author = Prelude Research
 author_email = support@preludesecurity.com
 description = For interacting with the Prelude API
